fixed the )+ bug

+++ src/compute/add/computations.c
@@ -51,7 +51,7 @@ static char *compute_sub(char *first, char *second, char *base)
         }
     }
     for (int i = 0; i < len + 2; i++)
-        result[i] = base[result[i]];
+        result[i] = base[(int) result[i]];
     //free(first);
     //free(second);
     return result;
@@ -97,8 +97,9 @@ char *my_add(char *first, char *second, char *base, char *ops)
         result[i + 1] -= result[i] * len_base;
     }
     for (int i = 0; i < len + 1; i++)
-        result[i] = base[result[i]];
+        result[i] = base[(int) result[i]];
     //free(first);
     //free(second);
     return result;
 }
+
